refactored contractor tenders list
